backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
PyMySQL
Werkzeug
python-dotenv
flask-jwt-extended
mysql-connector-python
python-dotenv
openai
scikit-learn
numpy
pandas
groq
